rename the OfTheDay atoms
